JMP-16 Remove ApiConstant class and update SecurityConfig; rename KeycloakRole to enums package; update Spring Boot version
