+++ build.gradle.kts
@@ -3,7 +3,7 @@ plugins {
     jacoco
     checkstyle
     id("application")
-    id("org.springframework.boot") version "3.4.3"
+    id("org.springframework.boot") version "3.4.4"
     id("io.spring.dependency-management") version "1.1.7"
 }
