+++ src/main/java/com/jmp/userservice/enums/KeycloakRole.java
@@ -1,4 +1,4 @@
-package com.jmp.userservice.constant;
+package com.jmp.userservice.enums;
 
 public enum KeycloakRole {
     ROLE_USER, ROLE_ADMIN, ROLE_MODERATOR, ROLE_SELLER, ROLE_SUPPORT
